annotation-pipeline-bak/requirements.txt
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
geopy>=2.3.0
googlemaps==4.10.0
geopy>=2.3.0
idna==3.4
importlib-resources==6.0.1
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
zipp==3.16.2
google-cloud-aiplatform
google-auth
google-auth-oauthlib
google-auth-httplib2